Rename ProcessingCLI to just CLI
I'm going to create a Processing namespace that contains a Sketch
and SketchDirectory class. All this will handle the creation,
naming, renaming, clonging, etc. of sketches. This stuff is
Processing biz logic, and therefore it should exist in the Processing
namespace. The invocation of this stuff is not, and should not be
included in the namespace.