Separate out Processing::SketchDirectory and Processing::Sketch
Makes CLI#create very lean, and that's the way I like it. Also, it
makes the test case for this method really simple.

Anything that involves sniffing out multiple sketch files is the
responsibility of the SketchDirectory. All that the Sketch class
does - currently - is handle saving. Later it will grow a bit to
handle renaming and cloning, too.