Open sketch file instead of sketch folder.
This makes for a nicer UX, as it means the user doesn't have to
manually open the only file in the sketch that's generated. That
file just opens on its own.

I also like it that the sketch is responsible for opening itself.
It'd be neat if I could push all File IO down to the sketch level.