Sanitize directory names before cloning.
When CLI is used in conjunction with command-line autocompletion
of directory names, an extra backslash is included with the source
name and passed into the clone operation. That caused problems
when it came to renaming the sketch file. This change sanitizes
the input before it's passed in, solving the problem.